Updated schematic and code
Added pulldown resistor instead of direct grounding on MOSI line due to not being able to program it.

Also added EEPROM code to store the calibration to not have to calibrate each time at startup. Calibration can be forced if the throttle is in the full on position when in the normal orientation